+++ resources/backend/js/gutenberg.js
@@ -885,6 +885,7 @@ function convertKitGutenbergRegisterBlock(block) {
 
 		// Register Block.
 		registerBlockType('convertkit/' + block.name, {
+			apiVersion: convertkit_gutenberg.block_api_version,
 			title: block.title,
 			description: block.description,
 			category: block.category,
@@ -971,7 +972,7 @@ function convertKitGutenbergRegisterPrePublishActions(actions) {
 				PluginPrePublishPanel,
 				{
 					className: 'convertkit-pre-publish-actions',
-					title: 'ConvertKit',
+					title: 'Kit',
 					initialOpen: true,
 				},
 				rows