+++ tests/EndToEnd/forms/post-types/PageFormCest.php
@@ -278,6 +278,9 @@ class PageFormCest
 
 		// Confirm no meta tag exists within the content.
 		$I->dontSeeInSource('<meta http-equiv="Content-Type" content="text/html; charset=utf-8">');
+
+		// Confirm no extra <html>, <head> or <body> tags are output i.e. injecting the form doesn't result in DOMDocument adding tags.
+		$I->seeNoExtraHtmlHeadBodyTagsOutput($I);
 	}
 
 	/**
@@ -324,6 +327,9 @@ class PageFormCest
 
 		// Confirm no meta tag exists within the content.
 		$I->dontSeeInSource('<meta http-equiv="Content-Type" content="text/html; charset=utf-8">');
+
+		// Confirm no extra <html>, <head> or <body> tags are output i.e. injecting the form doesn't result in DOMDocument adding tags.
+		$I->seeNoExtraHtmlHeadBodyTagsOutput($I);
 	}
 
 	/**
@@ -375,6 +381,9 @@ class PageFormCest
 
 		// Confirm no meta tag exists within the content.
 		$I->dontSeeInSource('<meta http-equiv="Content-Type" content="text/html; charset=utf-8">');
+
+		// Confirm no extra <html>, <head> or <body> tags are output i.e. injecting the form doesn't result in DOMDocument adding tags.
+		$I->seeNoExtraHtmlHeadBodyTagsOutput($I);
 	}
 
 	/**
@@ -427,8 +436,8 @@ class PageFormCest
 		// Confirm no meta tag exists within the content.
 		$I->dontSeeInSource('<meta http-equiv="Content-Type" content="text/html; charset=utf-8">');
 
-		// Confirm no meta tag exists within the content.
-		$I->dontSeeInSource('<meta http-equiv="Content-Type" content="text/html; charset=utf-8">');
+		// Confirm no extra <html>, <head> or <body> tags are output i.e. injecting the form doesn't result in DOMDocument adding tags.
+		$I->seeNoExtraHtmlHeadBodyTagsOutput($I);
 	}
 
 	/**
@@ -479,6 +488,9 @@ class PageFormCest
 
 		// Confirm no meta tag exists within the content.
 		$I->dontSeeInSource('<meta http-equiv="Content-Type" content="text/html; charset=utf-8">');
+
+		// Confirm no extra <html>, <head> or <body> tags are output i.e. injecting the form doesn't result in DOMDocument adding tags.
+		$I->seeNoExtraHtmlHeadBodyTagsOutput($I);
 	}
 
 	/**